Buff系统：持续型Buff测试通过，开始测试事件监听Buff

+++ Unity/Assets/Model/NKGMOBA/Buff/Core/BuffSystemBase.cs
@@ -23,7 +23,7 @@ namespace ETModel
         /// <summary>
         /// 最多持续到什么时候
         /// </summary>
-        public float MaxLimitTime;
+        public long MaxLimitTime;
         
         /// <summary>
         /// Buff数据
